requirements.txt
python-telegram-bot==20.7
openai==1.12.0
python-dotenv==1.0.0
SpeechRecognition==3.10.0
gTTS==2.5.0
pydub==0.25.1 